Primeira versão do Sala do Futuro (Login, Data [Geral e Agenda])

+++ auto_back/project/tests/test_json_handler.py
@@ -5,19 +5,18 @@ from project.utils import JSONHandler
 
 class TestJSONHandler(unittest.TestCase):
     def setUp(self):
-        self.test_file = "test_data.json"
+        self.test_file = 'test_data.json'
         self.json_handler = JSONHandler(self.test_file)
 
     def tearDown(self):
-        if os.path.exists(self.test_file):
-            os.remove(self.test_file)
+        if os.path.exists(self.test_file): os.remove(self.test_file)
 
     def test_carregar_dados_arquivo_inexistente(self):
         dados = self.json_handler.carregar()
         self.assertEqual(dados, {})
 
     def test_salvar_e_carregar_dados(self):
-        dados_teste = {"chave": "valor"}
+        dados_teste = {'chave': 'valor'}
         self.json_handler.salvar(dados_teste)
         dados_carregados = self.json_handler.carregar()
         self.assertEqual(dados_carregados, dados_teste)
@@ -27,5 +26,4 @@ class TestJSONHandler(unittest.TestCase):
         resultado = self.json_handler.salvar(dados_invalidos)
         self.assertFalse(resultado)
 
-if __name__ == "__main__":
-    unittest.main()
+if __name__ == '__main__': unittest.main()
